Add: Voltage & Temperature Logics
Not yet done

+++ lib/BQ76940/RegistersOV_UV_TRIP.hpp
@@ -4,8 +4,8 @@
 #include "RegistersClass.hpp"
 #include <stdint.h>
 
-extern Register OV_TRIP = Register(static_cast<uint8_t>(RegisterAddress::OV_TRIP), 0xAC);
-extern Register UV_TRIP = Register(static_cast<uint8_t>(RegisterAddress::UV_TRIP), 0x97);
+extern Register OV_TRIP = Register(RegisterAddress::OV_TRIP, 0xAC);
+extern Register UV_TRIP = Register(RegisterAddress::UV_TRIP, 0x97);
 
 void loadOVTrip();
 void loadUVTrip();
